Return argv when valid
